Allow use of gripper in both modes
