+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "capra_web_ui",
-  "version": "2.4.8",
+  "version": "2.4.9",
   "author": {
     "name": "Club Capra",
     "email": "admin@clubcapra.com",
