+++ src/renderer/inputSystem/flipperModeActions.ts
@@ -123,7 +123,7 @@ export const flipperModeActions: Action[] = [
       rosClient.callService({ name: '/headlights' }).catch(log.error)
     },
   },
-  {
+  /*{
     name: 'flipper_reset',
     bindings: [{ type: 'gamepadBtnDown', button: buttonMappings.B }],
     perform: () => {
@@ -131,7 +131,7 @@ export const flipperModeActions: Action[] = [
         .callService({ name: 'markhor/flippers/flipper_reset' })
         .catch(log.error)
     },
-  },
+  },*/
   {
     name: 'gamepad',
     bindings: [{ type: 'gamepad' }],
